pyproject.toml
[project]
name = "aorus-ai-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "markupsafe==2.1.5",
    "jinja2",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "torchaudio>=2.5.1",
    "llama-cpp-python>=0.3.16",
    "bs4>=0.0.2",
    "requests>=2.32.5",
    "playwright>=1.58.0",
    "faiss-cpu>=1.13.2",
    "sentence-transformers>=5.2.3",
    "opencc>=1.2.0",
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }
torchaudio = { index = "pytorch-cu121" }
